Add ReservationsService Tests

+++ Services/THECinema.Services.Data/ReservationsService.cs
@@ -13,36 +13,33 @@
     using THECinema.Services.Data.Contracts;
     using THECinema.Web.ViewModels.Halls;
     using THECinema.Web.ViewModels.Movies;
+    using THECinema.Web.ViewModels.Projections;
     using THECinema.Web.ViewModels.Reservations;
 
     public class ReservationsService : IReservationsService
     {
-        private readonly IDeletableEntityRepository<Projection> projectionsRepository;
+        private const string InvalidIdExceptionMessage = "Reservation doesn't exist!";
+
+        private readonly IDeletableEntityRepository<Reservation> reservationsRepository;
+        private readonly IProjectionsService projectionsService;
         private readonly IHallsService hallsService;
         private readonly IMoviesService moviesService;
         private readonly IDeletableEntityRepository<ProjectionSeat> seatsRepository;
-        private readonly IDeletableEntityRepository<Reservation> reservationsRepository;
-        private readonly IDeletableEntityRepository<Movie> moviesRepository;
-        private readonly IRepository<Hall> hallRepository;
         private readonly UserManager<ApplicationUser> userManager;
 
         public ReservationsService(
-            IDeletableEntityRepository<Projection> projectionsRepository,
+            IDeletableEntityRepository<Reservation> reservationsRepository,
+            IProjectionsService projectionsService,
             IHallsService hallsService,
             IMoviesService moviesService,
             IDeletableEntityRepository<ProjectionSeat> seatsRepository,
-            IDeletableEntityRepository<Reservation> reservationsRepository,
-            IDeletableEntityRepository<Movie> moviesRepository,
-            IRepository<Hall> hallRepository,
             UserManager<ApplicationUser> userManager)
         {
-            this.projectionsRepository = projectionsRepository;
+            this.reservationsRepository = reservationsRepository;
+            this.projectionsService = projectionsService;
             this.hallsService = hallsService;
             this.moviesService = moviesService;
             this.seatsRepository = seatsRepository;
-            this.reservationsRepository = reservationsRepository;
-            this.moviesRepository = moviesRepository;
-            this.hallRepository = hallRepository;
             this.userManager = userManager;
         }
 
@@ -66,6 +63,12 @@
         public async Task DeleteAsync(string id)
         {
             var reservation = this.reservationsRepository.All().Where(r => r.Id == id).FirstOrDefault();
+
+            if (reservation == null)
+            {
+                throw new ArgumentNullException(InvalidIdExceptionMessage);
+            }
+
             this.reservationsRepository.Delete(reservation);
             await this.reservationsRepository.SaveChangesAsync();
         }
@@ -92,9 +95,9 @@
                 return null;
             }
 
-            var projection = await this.projectionsRepository.GetByIdWithDeletedAsync(reservation.ProjectionId);
-            var movie = await this.moviesRepository.GetByIdWithDeletedAsync(projection.MovieId);
-            var hall = this.hallRepository.All().Where(h => h.Id == projection.HallId).FirstOrDefault();
+            var projection = this.projectionsService.GetByProjectionId<ProjectionViewModel>(reservation.ProjectionId);
+            var movie = this.moviesService.GetById<FullInfoMovieViewModel>(projection.MovieId);
+            var hall = this.hallsService.GetById<ReservationHallViewModel>(projection.HallId);
             var user = await this.userManager.FindByIdAsync(reservation.ApplicationUserId);
 
             var viewModel = new FullInfoReservationViewModel
@@ -112,9 +115,15 @@
             return viewModel;
         }
 
-        public async Task<ReservationViewModel> GetDetailsAsync(string projectionId)
+        public ReservationViewModel GetDetails(string projectionId)
        {
-            var projection = await this.projectionsRepository.GetByIdWithDeletedAsync(projectionId);
+            var projection = this.projectionsService.GetByProjectionId<ProjectionViewModel>(projectionId);
+
+            if (projection == null)
+            {
+                return null;
+            }
+
             var movie = this.moviesService.GetById<ReservationMovieViewModel>(projection.MovieId);
             var hall = this.hallsService.GetById<ReservationHallViewModel>(projection.HallId);
             var seats = new Dictionary<string, List<ProjectionSeat>>();
@@ -205,7 +214,7 @@
         public ParseReservationDataModel ParseData(string priceInput, string seatsInput)
         {
             var priceAsString = priceInput;
-            priceAsString = priceAsString.Substring(1, priceAsString.Length - 1);
+            priceAsString = priceAsString[1..];
             var price = double.Parse(priceAsString);
 
             var selectedSeatsArray = seatsInput.Split(",");
